Added Application viewer and Employer panel Modified JobController to handle status changes and application viewing Created factories for Job and Application

+++ app/Models/Application.php
@@ -15,10 +15,13 @@ class Application extends Model
         parent::boot();
 
         static::updated(function ($application) {
-            Log::info('Application updated');
             if ($application->isDirty('status')) {
-                Log::info('Application status changed');
-                $application->files()->detach();
+                if ($application->status === 'hired') {
+                    $application->files()->detach();
+                }
+                if ($application->status === 'closed') {
+                    $application->files()->detach();
+                }
             }
         });
     }
